+++ src/ingameWindows/iwPlayReplay.h
@@ -25,12 +25,11 @@
 class iwPlayReplay : public IngameWindow
 {
     public:
-        /// Konstruktor von @p iwPlayReplay.
         iwPlayReplay();
 
         void Msg_ButtonClick(const unsigned int ctrl_id) override;
         void Msg_MsgBoxResult(const unsigned msgbox_id, const MsgboxResult mbr) override;
-        void Msg_TableChooseItem(const unsigned ctrl_id, const unsigned short selection) override;
+        void Msg_TableChooseItem(const unsigned ctrl_id, const unsigned selection) override;
 
     private:
         void PopulateTable();
