+++ src/ingameWindows/iwStatistics.h
@@ -29,7 +29,6 @@ class iwStatistics : public IngameWindow
 
     public:
 
-        /// Konstruktor von @p iwStatistics.
         iwStatistics();
         ~iwStatistics() override;
 
@@ -46,7 +45,7 @@ class iwStatistics : public IngameWindow
 
         void Msg_ButtonClick(const unsigned int ctrl_id) override;
         void Msg_PaintAfter() override;
-        void Msg_OptionGroupChange(const unsigned int ctrl_id, const unsigned short selection) override;
+        void Msg_OptionGroupChange(const unsigned int ctrl_id, const int selection) override;
         void DrawStatistic(StatisticType type);
         void DrawAxis();
 };
