+++ src/ingameWindows/iwTrade.cpp
@@ -34,14 +34,6 @@
 // Include last!
 #include "DebugNew.h" // IWYU pragma: keep
 
-///////////////////////////////////////////////////////////////////////////////
-/**
- *  Konstruktor von @p iwBuilding.
- *
- *  @todo überprüfen und die restlichen Steuerelemente zur Funktion bringen
- *
- *  @author OLiver
- */
 iwTrade::iwTrade(nobBaseWarehouse& wh)
     : IngameWindow(wh.CreateGUIID(), (unsigned short) - 2, (unsigned short) - 2, 400, 194, _("Trade"), LOADER.GetImageN("resource", 41)),
       wh(wh), possibleSrcWarehouses(GAMECLIENT.GetLocalPlayer().GetWarehousesForTrading(wh))
@@ -126,7 +118,7 @@ void iwTrade::Msg_ButtonClick(const unsigned int  /*ctrl_id*/)
     }
 }
 
-void iwTrade::Msg_ComboSelectItem(const unsigned ctrl_id, const unsigned short selection)
+void iwTrade::Msg_ComboSelectItem(const unsigned ctrl_id, const int selection)
 {
     switch(ctrl_id)
     {
